+++ packages/core-web/src/PageModel.ts
@@ -17,11 +17,11 @@ export interface PageModel {
 export type ValueType = number | string
 
 function floatstr(value: string): boolean {
-  return /(\d+\.\d+)/g.test(value)
+  return /^(\d+\.\d+)$/g.test(value)
 }
 
 function numstr(value: string): boolean {
-  return /\d+/g.test(value)
+  return /^\d+$/m.test(value)
 }
 
 function getValue<T extends ValueType>(value: string): T {