+++ packages/core-web/src/middleware/LoggerMiddleware.ts
@@ -5,7 +5,7 @@ export function LoggerMiddleware(logger: Lincoln): RequestHandler {
   const log = logger.extend('logger')
 
   return (req: Request, res: Response, next: NextFunction) => {
-    log.debug(req.url, req.method, req.params, req.query)
+    log.debug(req.url, req.method, req.params, req.query, req.body)
     next()
   }
 }